Moved files out of /Assets/
