+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/Testers/ConnectionStateTester.cs
@@ -1,0 +1,22 @@
+﻿using UnityEngine;
+using UnityEngine.UI;
+
+namespace Fps
+{
+    public class ConnectionStateTester : MonoBehaviour
+    {
+        private Text text;
+
+        private void Awake()
+        {
+            ConnectionStateReporter.OnConnectionStateChange += StateChanged;
+            text = GetComponent<Text>();
+        }
+
+        private void StateChanged(ConnectionStateReporter.State state, string information)
+        {
+            text.text = $"State: {state.ToString()}\n" +
+                $"Info: {information}";
+        }
+    }
+}
